Use pre-created roles for issuing tokens
Creating policy per node at request time, which is
problematic due to the following:
* Orphaned tokens cannot be renewed indefinitely
* Token role lists cannot contain wildcards, thus the role(s)
  we use for creating tokens would need to be pre-generated or updated
  externally to enumerate all node policies.
* If this api were compromised, all secrets stored in vault could be
  compromised

This commit replaces orphaned tokens created using per-node policy with
tokens generated from a role corresponding to the system role assigned
to the node.  For example, if a node is a 'worker' in the 'bar' system,
it's token will be created using the 'credmanager-bar-worker' token
role.  Only the policies requested by the client will be applied to the
token.  If the client requests a policy that's not allowed by the role,
the token request will be rejected.