Refactor to issue secrets for approle auth
Using app role authentication, the credmanager api can be granted access
to issue single use secret_ids that when combined on the node with a
generated role_id unique to the client can be used to obtain a token.

This has some advantages:
1. The credmanager-api no longer needs to be able to issue tokens
directly
2. The policies for each client can be assigned to the app role on the
vault server, eliminating the need for credmanager-api to manage these.
3. If the role id is properly chosen and secured, it should improve the
security of the overall system.