Update for use without inventory backend, also update vault kv calls to support v2
Vault updated the default kv api to be v2.  This enables versioning and
moves the latest version read/write endpoints to secret/data/foo instead
of secret/foo.  It also modifies the data structure that is returned.

Removing the inventory backend as it adds complexity without adding a
ton of value.